refactor doc set and user group (mostly)
